+++ src/StefanoTree/NestedSet/Adapter/Zend1.php
@@ -449,7 +449,7 @@ class Zend1 extends AdapterAbstract implements AdapterInterface
     /**
      * {@inheritdoc}
      */
-    public function getPath($nodeId, int $startLevel = 0, bool $excludeLastNode = false): array
+    public function getAncestors($nodeId, int $startLevel = 0, bool $excludeLastNode = false): array
     {
         $options = $this->getOptions();