+++ src/StefanoTree/NestedSet/Adapter/Zend2.php
@@ -500,7 +500,7 @@ class Zend2 extends AdapterAbstract implements AdapterInterface
     /**
      * {@inheritdoc}
      */
-    public function getPath($nodeId, int $startLevel = 0, bool $excludeLastNode = false): array
+    public function getAncestors($nodeId, int $startLevel = 0, bool $excludeLastNode = false): array
     {
         $options = $this->getOptions();